--- src/components/main/About.jsx
@@ -1,4 +1,0 @@
-const AboutUs = () => {
-  return <div>About-us</div>;
-};
-export default AboutUs;
