--- src/components/main/ClientReviews.jsx
@@ -1,4 +1,0 @@
-const ClientReviews = () => {
-  return <div>ClientReviews</div>;
-};
-export default ClientReviews;
