--- src/components/main/ContactUs.jsx
@@ -1,4 +1,0 @@
-const ContactUs = () => {
-  return <div>ContactUs</div>;
-};
-export default ContactUs;
