--- src/components/main/Footer.jsx
@@ -1,4 +1,0 @@
-const Footer = () => {
-  return <div>Footer</div>;
-};
-export default Footer;
