--- src/components/main/Services.jsx
@@ -1,4 +1,0 @@
-const Services = () => {
-  return <div>Services</div>;
-};
-export default Services;
